added role linking

+++ Icarus/Services/GoIService.cs
@@ -1,5 +1,7 @@
 ﻿using Icarus.Context;
 using Icarus.Context.Models;
+using Icarus.Exceptions;
+using Icarus.Utils;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -13,11 +15,32 @@ namespace Icarus.Services
     {
         private readonly CharacterService _characterService;
         private readonly RoleService _roleService;
+        private readonly TickService _tickService;
 
-        public GoIService(CharacterService characterService, RoleService roleService)
+        public GoIService(CharacterService characterService, RoleService roleService, TickService tickService)
         {
             _characterService = characterService;
             _roleService = roleService;
+            _tickService = tickService;
+
+            _tickService.TickCheckEvent += SyncAllGoiRoles;
+        }
+
+        public async Task LinkGoiRole(int goiId, ulong roleId)
+        {
+            using var db = new IcarusContext();
+
+            var goi = db.GroupOfInterests.SingleOrDefault(g => g.Id == goiId);
+
+            if (goi == null) 
+            {
+                throw new GroupOfInterestNotFoundException();
+            }
+
+            goi.DiscordRoleId = roleId;
+
+            db.Update(goi);
+            await db.SaveChangesAsync();
         }
 
         public async Task<string> AddGoI(string name)
@@ -55,6 +78,18 @@ namespace Icarus.Services
             return await db.GroupOfInterests.ToListAsync();
         }
 
+        public async void SyncAllGoiRoles()
+        {
+            var settings = ConfigFactory.GetConfig();
+
+            var allLivingCharacters = await _characterService.GetAllCharactersIncludeGoI();
+
+            foreach (var character in allLivingCharacters)
+            {
+                _ = SyncGoiRoles(ulong.Parse(character.DiscordUserId), settings.GuildId);
+            }
+        }
+
         public async Task SyncGoiRoles(ulong discordId, ulong guildId)
         {
             using var db = new IcarusContext();
@@ -63,9 +98,14 @@ namespace Icarus.Services
 
             var desiredRoleId = character.GroupOfInterest.DiscordRoleId;
 
+            if (desiredRoleId == null) 
+            {
+                return;
+            }
+
             var allGroups = await GetAllGroups();
 
-            var unwantedRoleIds = allGroups.Select(ag => ag.DiscordRoleId).ToList();
+            var unwantedRoleIds = allGroups.Where(ag => ag.DiscordRoleId != null).Select(ag => ag.DiscordRoleId).ToList();
             unwantedRoleIds.Remove(desiredRoleId);
 
             await _roleService.RemoveRoles(unwantedRoleIds, discordId, guildId);
